Fix reconciler config
