+++ controllers/operator.marin3r/discoveryservice_controller.go
@@ -36,18 +36,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-func init() {
-	reconciler.Config.AnnotationsDomain = "marin3r.3scale.net"
-	reconciler.Config.ResourcePruner = false
-	reconciler.Config.ManagedTypes = reconciler.NewManagedTypes().
-		Register(&appsv1.DeploymentList{}).
-		Register(&corev1.ServiceList{}).
-		Register(&rbacv1.RoleList{}).
-		Register(&rbacv1.RoleBindingList{}).
-		Register(&corev1.ServiceAccountList{}).
-		Register(&operatorv1alpha1.DiscoveryServiceCertificateList{})
-}
-
 // DiscoveryServiceReconciler reconciles a DiscoveryService object
 type DiscoveryServiceReconciler struct {
 	reconciler.Reconciler
